refactor: component에 있는 파일에 목적에 안맞는 파일 screen 폴더로 이동, pages폴더 screen으로 이름 변경.
