+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import './App.css';
 import Router from './Router';
-import Sidebar from './components/sidebar/Sidebar';
+import Sidebar from './screen/sidebar/Sidebar';
 import Modal from 'react-modal';
 import { ToastContainer } from 'react-toastify';
 Modal.setAppElement('#root');
