+++ src/Router.js
@@ -1,16 +1,14 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import SignUp from '~/components/signup/SignUp';
-import DashBoard from '~/pages/DashBoard';
-import GenerateRecommendName from '~/pages/GenerateRecommendName/GenerateRecommendName';
-import RecommendNameScreen from '~/pages/RecommendNameScreen/RecommendNameScreen';
+import GenerateRecommendNameScreen from '~/screen/GenerateRecommendNameScreen/GenerateRecommendNameScreen';
+import RecommendNameScreen from '~/screen/RecommendNameScreen/RecommendNameScreen';
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<GenerateRecommendName />} />
-        <Route path="/signup" element={<SignUp />} />
-        <Route path="/dashboard" element={<DashBoard />} />
+        <Route path="/" element={<GenerateRecommendNameScreen />} />
         <Route path="/result" element={<RecommendNameScreen />} />
+        {/* <Route path="/signup" element={<SignUp />} /> */}
       </Routes>
     </BrowserRouter>
   );
