+++ src/screen/GenerateRecommendNameScreen/GenerateRecommendNameScreen.css
