+++ src/screen/RecommendNameScreen/RecommendName/RecommendNameCard/RecommendNameCard.css
