+++ src/screen/RecommendNameScreen/RecommendName/RecommendNameCard/RecommendNameCard.tsx
