+++ src/screen/RecommendNameScreen/RecommendNameSetting/RecommendNameSetting.css
