+++ src/screen/RecommendNameScreen/RecommendNameSetting/SettingOptionCard/SettingOptionCard.css
