+++ src/screen/RecommendNameScreen/RecommendNameSetting/SettingOptionCard/SettingOptionCard.tsx
